Move `DummyLocation` to `ValueObjects` namespace
