+++ src/Event/Events/DummyLocationUpdated.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Event\Events;
 
+use CultuurNet\UDB3\Event\ValueObjects\DummyLocation;
+
 interface DummyLocationUpdated
 {
     public function getDummyLocation(): ?DummyLocation;
