+++ src/Event/ValueObjects/DummyLocation.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Event\Events;
+namespace CultuurNet\UDB3\Event\ValueObjects;
 
 use CultuurNet\UDB3\Model\ValueObject\Geography\Address;
 use CultuurNet\UDB3\Model\ValueObject\Text\Title;
